% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-bibliometrix_parsing.R
\name{bibliometrix_parsing}
\alias{bibliometrix_parsing}
\title{Wrapper function to call bibliometrix's convert2df() for a given set of records}
\usage{
bibliometrix_parsing(dataset_list, n_threads = parallel::detectCores())
}
\arguments{
\item{dataset_list}{\itemize{
\item a list of the files in each dataset. Each is a list containing 3 elements: \emph{file} (either a +-pç34yghtsingle file or vector with multiple files containing bibliographic records), \emph{dbsource} (bibliographic source), and \emph{format} (file format)
Example: datasets <- list(dimensions = list(file = 'path/to/dimensions.csv', dbsource = 'dimensions', format = 'csv'), scopus = list(file = c('path/to/scopus/1.bib', 'path/to/scopus/2.bib'), dbsource = 'scopus', format = 'bibtex'), wos = list(file = c('path/to/wos/1.bib','path/to/wos/2.bib'), dbsource = 'wos', format = 'bibtex'), 'lens' = list(file = 'path/to/lens.csv', dbsource = 'lens', format = 'csv') )
}}

\item{n_threads}{\itemize{
\item number of (logical) cores used in the parsing process. By default, uses the number of cores dectected by parallel:detectCores(). Internally, however, this parameter will not exceed the number datasets provided, since going above this threshold would not increase performance.
}}

\item{db_data}{\itemize{
\item list containing 3 elements: file (either a single file or vector with multiple files containing bibliographic records), dbsource (bibliographic source), and format (file format)
}}
}
\value{
a database featuring all records from the files specified, formatted according to bibliometrix fields

a db_list
}
\description{
Wrapper function to call bibliometrix's convert2df() for a given set of records
}
\seealso{
\code{\link[bibliometrix:convert2df]{bibliometrix::convert2df()}}

Function to convert list of database files into db_lists using bibliometrix
}
