% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliover_app.R
\name{biblioverApp}
\alias{biblioverApp}
\title{Title}
\usage{
biblioverApp(port = NULL, max_upload_size = 1000, launch.browser = TRUE)
}
\arguments{
\item{port}{- port of the application}

\item{max_upload_size}{- max upload size of documents (in MB) - Default 100}

\item{launch.browser}{- launch on browser - Default = TRUE}
}
\value{
opens a instance of the biblioverlap UI
}
\description{
Title
}
