% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-biblioverlap.R
\name{biblioverlap}
\alias{biblioverlap}
\title{Obtains document overlap between databases from a named list}
\usage{
biblioverlap(
  db_list,
  db_order = names(db_list),
  matching_fields = default_matching_fields,
  n_threads = parallel::detectCores(),
  ti_penalty = 0.1,
  ti_max = 0.6,
  so_penalty = 0.1,
  so_max = 0.3,
  au_penalty = 0.1,
  au_max = 0.3,
  py_max = 0.3,
  score_cutoff = 1
)
}
\arguments{
\item{db_list}{\itemize{
\item list of dataframes containing the sets of bibliographic data
}}

\item{db_order}{\itemize{
\item order of the databases
}}

\item{matching_fields}{\itemize{
\item Column names used in the matching
}}

\item{n_threads}{\itemize{
\item number of (logical) cores used in the matching procedures
}}

\item{ti_penalty}{\itemize{
\item penalty applied for each increment in Title's levenshtein distance
}}

\item{ti_max}{\itemize{
\item max score value for Title
}}

\item{so_penalty}{\itemize{
\item penalty applied for each increment in Source's levenshtein distance
}}

\item{so_max}{\itemize{
\item max score value for Source
}}

\item{au_penalty}{\itemize{
\item penalty applied for each increment in Author's levenshtein distance
}}

\item{au_max}{\itemize{
\item max  score value for Author
}}

\item{py_max}{\itemize{
\item max score value for Publication Year
}}

\item{score_cutoff}{\itemize{
\item minimum final score for a valid match between two documents
}}
}
\value{
a modified version of db_list where matching documents share the same UUID
}
\description{
Obtains document overlap between databases from a named list
}
