% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-score_matching.R
\name{calc_exact_score_matrix}
\alias{calc_exact_score_matrix}
\title{Calculates score matrix for a pair of databases based on exact match}
\usage{
calc_exact_score_matrix(db1_col, db2_col, max_score)
}
\arguments{
\item{db1_col}{- Column from db1 with values to be compared}

\item{db2_col}{- Column from db2 with values to be compared}

\item{max_score}{- Max score value}
}
\value{
a sparseMatrix object containing score values
}
\description{
Calculates score matrix for a pair of databases based on exact match
}
