% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-bibliometrix_parsing.R
\name{convert2df_wrapper}
\alias{convert2df_wrapper}
\title{Wrapper function to call bibliometrix's convert2df() for a given set of records}
\usage{
convert2df_wrapper(db_data)
}
\arguments{
\item{db_data}{- list containing 3 elements: file (either a single file or vector with multiple files containing bibliographic records), dbsource (bibliographic source), and format (file format)}
}
\value{
a database featuring all records from the files specified, formatted according to bibliometrix fields
}
\description{
Wrapper function to call bibliometrix's convert2df() for a given set of records
}
\seealso{
[bibliometrix::convert2df()]
}
