% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_matching.R
\name{convert_to_sparse_matrix}
\alias{convert_to_sparse_matrix}
\title{Converts a matrix to sparseMatrix format}
\usage{
convert_to_sparse_matrix(matrix)
}
\arguments{
\item{matrix}{- A matrix object}
}
\value{
a matrix in sparseMatrix format, where negative values have been replaced by 0
}
\description{
This is a helper function to both calc_distance_score_matrix() and calc_exact_score_matrix()
}
