% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_preprocessing.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Processes bibliographic data for the downstream matching procedure}
\usage{
data_preprocessing(db_list, matching_fields)
}
\arguments{
\item{db_list}{\itemize{
\item list containing a dataframe for each set
}}

\item{matching_fields}{\itemize{
\item list containing the names of the columns used in the matching procedure
}}
}
\value{
a list containing a (modified) dataframe for each set
}
\description{
Processes bibliographic data for the downstream matching procedure
}
