% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{db_analyses}
\alias{db_analyses}
\title{Obtains document overlap between databases from a named list}
\usage{
db_analyses(
  db_list,
  db_order = names(db_list),
  n_threads = parallel::detectCores(),
  ti_penalty = 0.1,
  ti_max = 0.6,
  j9_penalty = 0.1,
  j9_max = 0.3,
  au_penalty = 0.1,
  au_max = 0.3,
  py_max = 0.3,
  score_cutoff = 1
)
}
\arguments{
\item{db_list}{- list of dataframes containing the sets of bibliographic data}

\item{db_order}{- order of the databases}

\item{n_threads}{- number of (logical) cores used in the matching procedures}

\item{ti_penalty}{- penalty applied for each increment in Title's levenshtein distance}

\item{ti_max}{- max score value for Title}

\item{j9_penalty}{- penalty applied for each increment in Source's levenshtein distance}

\item{j9_max}{- max score value for Source}

\item{au_penalty}{- penalty applied for each increment in Author's levenshtein distance}

\item{au_max}{- max  score value for Author}

\item{py_max}{- max score value for Publication Year}

\item{score_cutoff}{- minimum final score for a valid match between two documents}
}
\value{
a modified version of db_list where matching documents share the same UUID
}
\description{
Obtains document overlap between databases from a named list
}
