% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{default_matching_fields}
\alias{default_matching_fields}
\title{Default matching fields for \code{\link[=biblioverlap]{biblioverlap()}}}
\format{
An object of class \code{list} of length 5.
}
\usage{
default_matching_fields
}
\description{
Each element in the named list corresponds to a column that will be used to match documents between bibliographic datasets.
}
\details{
In total: there are five columns that need to be specified:
}
\section{Fields}{

\describe{
\item{\code{doi}}{Unique identifier (e.g. DOI)}

\item{\code{title}}{Document title}

\item{\code{pubyear}}{Publication year}

\item{\code{authors}}{Authors}

\item{\code{source}}{Publication source}
}}

\keyword{datasets}
