% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-doi_matching.R
\name{doi_matching}
\alias{doi_matching}
\title{Document matching using DOI (Digital Object Identifier) data for two dataframes}
\usage{
doi_matching(db1, db2, n_threads, doi_score = 2)
}
\arguments{
\item{db1}{\itemize{
\item First bibliographic database in the comparison
}}

\item{db2}{\itemize{
\item Second bibliographic database in the comparison
}}

\item{n_threads}{\itemize{
\item number of (logical) cores to be used in the matching
}}

\item{doi_score}{\itemize{
\item custom score value to help identify DOI matches in previous comparisons
}}
}
\value{
a list containing match data
}
\description{
Document matching using DOI (Digital Object Identifier) data for two dataframes
}
