% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-score_matching.R
\name{extract_score_matches}
\alias{extract_score_matches}
\title{Get the max score value for each row of the final score matrix}
\usage{
extract_score_matches(final_score_matrix, db1_index, db2_index)
}
\arguments{
\item{final_score_matrix}{- matrix with final score values}

\item{db1_index}{- db1 index column}

\item{db2_index}{- db2 index column}
}
\value{
a list containing match data
}
\description{
Get the max score value for each row of the final score matrix
}
