% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_preprocessing.R
\name{generating_db_col}
\alias{generating_db_col}
\title{Title}
\usage{
generating_db_col(db_list, matching_fields)
}
\arguments{
\item{db_list}{\itemize{
\item blabla
}}

\item{matching_fields}{-blabla}
}
\value{
blabla
}
\description{
Title
}
