% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-biblioverlap.R
\name{get_all_subset_matches}
\alias{get_all_subset_matches}
\title{Get all matches from a given subset of biblioverlap's results}
\usage{
get_all_subset_matches(subset_db_list, db_list)
}
\arguments{
\item{subset_db_list}{\itemize{
\item a subset of the results generated by \code{\link{biblioverlap}}
}}

\item{db_list}{\itemize{
\item the full set of results generated by \code{\link{biblioverlap}}
}}
}
\value{
\itemize{
\item the subset data plus any other records outside the subset that have been matched to its documents
}
}
\description{
Get all matches from a given subset of biblioverlap's results
}
\examples{
#Running document-level matching procedure for two datasets
biblioverlap_results <- biblioverlap(ufrj_bio_0122[1:2])$db_list

#Change the document type of one of the datasets from 'journal article' to
#'article' to emulate bibliographical source differences
biblioverlap_results[[2]][['Publication Type']] <- gsub('journal article',
'article',
biblioverlap_results[[2]][['Publication Type']])

#Generating venn diagram for the entire dataset
venn_plot(biblioverlap_results)

#Obtaining only the subset of records with publication type 'article'
biblioverlap_results_subset <- lapply(biblioverlap_results, function(db) {
db[db$'Publication Type' == "article", ] })

#Generating venn diagram for data subset
#Returns us how many documents categorized as 'article' are unique to a given
#dataset and how many find a match against other documents in the subset
#(i.e. that are also categorized as 'article', in this example)
venn_plot(biblioverlap_results_subset)

#Recovering missing matches due to bibliographical source differences
#in the subsetting process
subset_all_matches <- get_all_subset_matches(biblioverlap_results_subset,
biblioverlap_results)

#Generating venn diagram for data subset plus all its matches
#Returns us how many documents categorized as 'article' are unique to a given
#dataset and how many find a match against any other document
venn_plot(subset_all_matches)
}
