% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-bibliometrix_parsing.R
\name{get_matching_summary}
\alias{get_matching_summary}
\title{Getting a matching summary (plot and dataframe) of biblioverlap matching results}
\usage{
get_matching_summary(matched_db_list)
}
\arguments{
\item{matched_db_list}{- list of matched dataframes (with UUID column added by biblioverlap)}
}
\value{
a list containing a summary (df and plot) of the matching results
}
\description{
Getting a matching summary (plot and dataframe) of biblioverlap matching results
}
