% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-score_matching.R
\name{get_max_score_matches}
\alias{get_max_score_matches}
\title{Get the max score value for each row of the final score matrix}
\usage{
get_max_score_matches(final_score_matrix)
}
\arguments{
\item{final_score_matrix}{- matrix with final score values}
}
\value{
a list containing match data
}
\description{
Get the max score value for each row of the final score matrix
}
