% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plots_and_app.R
\name{get_uuid_list}
\alias{get_uuid_list}
\title{Extracting list of uuid columns from biblioverlap results}
\usage{
get_uuid_list(db_list)
}
\arguments{
\item{db_list}{\itemize{
\item list of matched dataframes (with UUID column added by biblioverlap)
}}
}
\value{
\itemize{
\item a list containing only the UUID data
}
}
\description{
Helper function to \code{\link[=plot_venn]{plot_venn()}} and \code{\link[=plot_upset]{plot_upset()}} that extracts overlapping data stored in the UUID column.
}
\keyword{internal}
