% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{global variables}
\alias{global variables}
\title{Variables for unquoted use in dplyr}
\description{
Internally, biblioverlap uses specific field names to refer to the data it processes. They are added here for unquoted use in dplyr.
}
\keyword{internal}
