% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-biblioverlap.R
\name{inherit_uuid_col}
\alias{inherit_uuid_col}
\title{Adding UUID column to original data provided by user}
\usage{
inherit_uuid_col(unique_db_list, internal_db_list)
}
\arguments{
\item{unique_db_list}{\itemize{
\item db_list after duplicate cleaning process by \code{\link[=removing_duplicates]{removing_duplicates()}}
}}

\item{internal_db_list}{\itemize{
\item internal_db_list used by \code{\link[=biblioverlap]{biblioverlap()}} for document matching
}}
}
\value{
a db_list with unique records that
}
\description{
This is a helper function to \code{\link[=biblioverlap]{biblioverlap()}} that extracts the UUID column from the internal data used for document matching and adds it to the original data provided by the user.
}
\keyword{internal}
