% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-biblioverlap.R
\docType{data}
\name{matching_fields}
\alias{matching_fields}
\title{My Custom List Object}
\format{
A named list.
}
\usage{
matching_fields
}
\description{
This is an example list object containing sample data.
}
\section{Fields}{

\describe{
\item{\code{doi}}{Description of the field.}

\item{\code{title}}{Description of the field.}

\item{\code{pubyear}}{Description of the field.}

\item{\code{authors}}{Description of the field.}

\item{\code{source}}{Description of the field.}
}}

\keyword{datasets}
