% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-bibliometrix_parsing.R
\name{matching_summary_df}
\alias{matching_summary_df}
\title{Obtaining a table summary of biblioverlap's matching}
\usage{
matching_summary_df(matched_db_list)
}
\arguments{
\item{matched_db_list}{- list of matched dataframes (with UUID column added by biblioverlap)}
}
\value{
a table summary the matching results
}
\description{
Obtaining a table summary of biblioverlap's matching
}
