% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-bibliometrix_parsing.R
\name{matching_summary_plot}
\alias{matching_summary_plot}
\title{Obtaining a plot summary of biblioverlap's matching}
\usage{
matching_summary_plot(matching_summary_df)
}
\arguments{
\item{matching_summary_df}{- summary of matching process generated by `matching_summary_df()`}
}
\value{
a barplot summary of the matching results
}
\description{
Obtaining a plot summary of biblioverlap's matching
}
