% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_preprocessing.R
\name{merge_input_files}
\alias{merge_input_files}
\title{Merge multiple input files from the same source}
\usage{
merge_input_files(input_files, sep = ",", quote = "\\"")
}
\arguments{
\item{input_files}{\itemize{
\item an array containing the path to all input files
}}

\item{sep}{\itemize{
\item field separator. Default: comma (',')
}}

\item{quote}{\itemize{
\item quote type used for character fields. Default: Double quotes ('"')
}}
}
\value{
a single dataframe with all unique records from the input files
}
\description{
Merge multiple input files from the same source
}
\details{
It is fairly common to retrieve data from a single bibliographic database in small chunks. Thus, this function is designed to merge multiple files from the same database into a single file while also removing duplicate records.
}
\examples{

## Generating tempfiles
tempfile1 <- tempfile(fileext = ".csv")
tempfile2 <- tempfile(fileext = ".csv")
write.csv(ufrj_bio_0122$Biochemistry, file = tempfile1, row.names = FALSE)
write.csv(ufrj_bio_0122$Genetics, file = tempfile2, row.names = FALSE)

## Testing function
merged_files <- merge_input_files(c(tempfile1, tempfile2))
dim(merged_files)
head(merged_files)

}
