% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-biblioverlap.R
\name{merge_results}
\alias{merge_results}
\title{Merge biblioverlap's results into a single dataframe}
\usage{
merge_results(db_list, filter = "none")
}
\arguments{
\item{db_list}{\itemize{
\item list of matched dataframes (with UUID column added by \code{\link{biblioverlap}})
}}

\item{filter}{\itemize{
\item value determining whether to return all the data ('none'), distinct records only ('distinct') or matched records only ('matched'). Default: 'none'
}}
}
\value{
a single dataframe containing data from db_list, featuring an additional 'SET_NAME' column to indicate from which dataset each record came
}
\description{
Merge biblioverlap's results into a single dataframe
}
\examples{

#Running document-level matching procedure for two datasets
biblioverlap_results <- biblioverlap(ufrj_bio_0122[1:2])

#Obtaining the results as a single dataframe (all records)
all_data <- merge_results(biblioverlap_results$db_list)

#Checking number of total rows and overlapping documents are in the dataframe
nrow(all_data)
sum(duplicated(all_data$UUID))

#Obtaining only distinct records as a single dataframe
distinct_data <- merge_results(biblioverlap_results$db_list, filter = 'distinct')

#Checking number of total rows and overlapping documents are in the dataframe
nrow(distinct_data)
sum(duplicated(distinct_data$UUID))

#Obtaining only matched records as a single dataframe
matched_data <- merge_results(biblioverlap_results$db_list, filter = 'matched')

#Checking number of total rows and overlapping documents are in the dataframe
nrow(matched_data)
sum(duplicated(matched_data$UUID))

}
