% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_matching.R
\name{parallel_doi_match}
\alias{parallel_doi_match}
\title{Document matching using DOI (Digital Object Identifier) data for two dataframes}
\usage{
parallel_doi_match(db1, db2, n_threads, doi_score = 2)
}
\arguments{
\item{db1}{- First bibliographic database in the comparison}

\item{db2}{- Second bibliographic database in the comparison}

\item{n_threads}{- number of (logical) cores to be used in the matching}

\item{doi_score}{- custom score value to help identify DOI matches in previous comparisons}
}
\value{
a list containing match data
}
\description{
Document matching using DOI (Digital Object Identifier) data for two dataframes
}
