% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plots_and_app.R
\name{plot_venn}
\alias{plot_venn}
\title{Plotting Venn Diagram from biblioverlap results}
\usage{
plot_venn(db_list, ...)
}
\arguments{
\item{db_list}{\itemize{
\item list of matched dataframes (with UUID column added by biblioverlap)
}}

\item{...}{\itemize{
\item Additional arguments that can be passed down to \code{\link[ggVennDiagram:ggVennDiagram]{ggVennDiagram::ggVennDiagram()}}
}}
}
\value{
a Venn Diagram
}
\description{
Plotting Venn Diagram from biblioverlap results
}
\examples{
#Running document-level matching procedure
matched_data <- biblioverlap(ufrj_bio_0122)

#Plotting the matching summary
plot_venn(matched_data$db_list)

}
