% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_preprocessing.R
\name{removing_duplicates}
\alias{removing_duplicates}
\title{Removes duplicates from each bibliographic dataset}
\usage{
removing_duplicates(db_list, matching_fields)
}
\arguments{
\item{db_list}{\itemize{
\item input list containing a dataframe for each set
}}

\item{matching_fields}{\itemize{
\item list containing the names of the columns to be used in the matching procedure
}}
}
\value{
a list containing a (duplicate-free) dataframe for each set
}
\description{
Removes duplicates from each bibliographic dataset
}
\keyword{internal}
