% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-score_matching.R
\name{score_matching}
\alias{score_matching}
\title{Calculating the final score matrix and extracting the highest scoring matches}
\usage{
score_matching(
  db1,
  db2,
  n_threads,
  ti_penalty,
  ti_max,
  so_penalty,
  so_max,
  au_penalty,
  au_max,
  py_max,
  score_cutoff
)
}
\arguments{
\item{db1}{\itemize{
\item First bibliographic database in the comparison
}}

\item{db2}{\itemize{
\item Second bibliographic database in the comparison
}}

\item{n_threads}{\itemize{
\item number of (logical) cores to be used in the matching
}}

\item{ti_penalty}{\itemize{
\item penalty applied for each increment in Title's levenshtein distance
}}

\item{ti_max}{\itemize{
\item max score value for Title
}}

\item{so_penalty}{\itemize{
\item penalty applied for each increment in Source's levenshtein distance
}}

\item{so_max}{\itemize{
\item max score value for Source
}}

\item{au_penalty}{\itemize{
\item penalty applied for each increment in Author's levenshtein distance
}}

\item{au_max}{\itemize{
\item max score value for Author
}}

\item{py_max}{\itemize{
\item max score value for Publication Year
}}

\item{score_cutoff}{\itemize{
\item minimum final score for a valid match between two documents
}}
}
\value{
a list containing match data
}
\description{
This function calculates each component of the score in a different score matrix, then sums those in order to obtain the final score.
Finally, it extracts the highest score match for each row of the final score matrix.
}
