% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-matching_helper_funcs.R
\name{subset_db_for_doi_match}
\alias{subset_db_for_doi_match}
\title{Subsetting bibliographic database records to use in the unique identifier matching procedure}
\usage{
subset_db_for_doi_match(db)
}
\arguments{
\item{db}{\itemize{
\item dataframe containing all fields from the bibliographic database records
}}
}
\value{
a subset of the database containing only relevant fields for the unique id matching procedure
}
\description{
Subsetting bibliographic database records to use in the unique identifier matching procedure
}
