% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-matching_helper_funcs.R
\name{subset_db_for_score_match}
\alias{subset_db_for_score_match}
\title{Subsetting bibliographic database records to use in the score matching procedure}
\usage{
subset_db_for_score_match(df)
}
\arguments{
\item{df}{\itemize{
\item a dataframe from a bibliographic database
}}
}
\value{
a subset df with all needed data for the score matching procedure
}
\description{
Subsetting bibliographic database records to use in the score matching procedure
}
