% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{test_data}
\alias{test_data}
\title{biobr22}
\format{
## `test_data`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
test_data
}
\description{
Data obtained from [The Lens Scholarly Search](https://www.lens.org)
Report ...
}
\keyword{datasets}
