% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-post_matching.R
\name{update_db2_matches}
\alias{update_db2_matches}
\title{Updating db2 matched records}
\usage{
update_db2_matches(db1, db2, match_list)
}
\arguments{
\item{db1}{- First bibliographic database in the comparison}

\item{db2}{- Second bibliographic database in the comparison}

\item{match_list}{- List containing all matching documents between db1 and db2}
}
\value{
db2 with updated uuid and score values
}
\description{
Updating db2 matched records
}
