% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plots_and_app.R
\name{upset_plot}
\alias{upset_plot}
\title{Plotting UpSet plot from biblioverlap results}
\usage{
upset_plot(db_list, add_logo = TRUE, ...)
}
\arguments{
\item{db_list}{\itemize{
\item list of matched dataframes (with UUID column added by \code{\link{biblioverlap}})
}}

\item{add_logo}{\itemize{
\item boolean specifying whether to add package logo to plot or not. Default: TRUE
}}

\item{...}{\itemize{
\item arguments to be passed down to \code{\link[UpSetR:upset]{UpSetR::upset}}
}}
}
\value{
a UpSet plot representation of document overlap between the input datasets
}
\description{
Plotting UpSet plot from biblioverlap results
}
\examples{
#Running document-level matching procedure
biblioverlap_results <- biblioverlap(ufrj_bio_0122[1:2])

#Checking biblioverlap results (db_list)
lapply(biblioverlap_results$db_list, head, n=1)

#Plotting the UpSet plot
upset_plot(biblioverlap_results$db_list)

}
