% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_empty_dfs.R
\name{any_empty_dfs}
\alias{any_empty_dfs}
\title{Take two dataframes and checks if any of them are empty}
\usage{
any_empty_dfs(df1, df2)
}
\arguments{
\item{df1}{- A data.frame object}

\item{df2}{- A second data.frame object}
}
\value{
a boolean value
}
\description{
This function takes a pair of dataframes as parameters and checks if any of them are empty (i.e. have 0 rows).

If at least one of the dataframes are empty, returns TRUE. Else, returns FALSE.

It is intended to help the document matching functions deal with datasets that have only DOI (or DOIless) records.
}
\examples{
any_empty_dfs(scopus_df, wos_df)
}
