% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-matching_helper_funcs.R
\name{any_empty_dfs}
\alias{any_empty_dfs}
\title{Take two dataframes and checks if any of them are empty}
\usage{
any_empty_dfs(df1, df2)
}
\arguments{
\item{df1}{\itemize{
\item A data.frame object
}}

\item{df2}{\itemize{
\item A second data.frame object
}}
}
\value{
a boolean value

df1 <- data.frame(A = sample(1:10)) #non-empty dataframe
df2 <- data.frame() #empty dataframe
df3 <- data.frame(A = c('test1', 'test2')) #non-empty dataframe
any_empty_dfs(df1, df2) #TRUE, as df2 is empty
any_empty_dfs(df1, df3) #FALSE, as there are no empty dfs in the arguments
}
\description{
This function takes a pair of dataframes as parameters and checks if any of them are empty (i.e. have 0 rows).

If at least one of the dataframes are empty, returns TRUE. Else, returns FALSE.

It is intended to help the document matching functions deal with datasets that have only DOI (or DOIless) records.
}
