% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plots_and_app.R
\name{biblioverApp}
\alias{biblioverApp}
\title{Shiny App for the biblioverlap package}
\usage{
biblioverApp(port = NULL, max_upload_size = 1000, launch.browser = TRUE)
}
\arguments{
\item{port}{\itemize{
\item port of the application
}}

\item{max_upload_size}{\itemize{
\item max upload size of documents (in MB) - Default 100
}}

\item{launch.browser}{\itemize{
\item launch on browser - Default = TRUE
}}
}
\value{
opens a instance of the biblioverlap UI
}
\description{
Shiny App for the biblioverlap package
}
\examples{
#Running the ShinyApp
\dontrun{biblioverApp()}

}
