% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-score_matching.R
\name{calc_distance_score_matrix}
\alias{calc_distance_score_matrix}
\title{Calculates score matrix for a pair of databases based on edit (levenshtein) distance}
\usage{
calc_distance_score_matrix(db1_col, db2_col, max_score, penalty, n_threads)
}
\arguments{
\item{db1_col}{- Column from db1 with values to be compared}

\item{db2_col}{- Column from db2 with values to be compared}

\item{max_score}{- Max score value (starting score value)}

\item{penalty}{- Penalty for each increase in edit distance}

\item{n_threads}{- Number of threads used for matrix calculation}
}
\value{
a sparseMatrix object containing score values
}
\description{
Calculates score matrix for a pair of databases based on edit (levenshtein) distance
}
