% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-matching_helper_funcs.R
\name{convert_to_sparse_matrix}
\alias{convert_to_sparse_matrix}
\title{Converts a regular matrix to sparseMatrix}
\usage{
convert_to_sparse_matrix(matrix)
}
\arguments{
\item{matrix}{\itemize{
\item A matrix object
}}
}
\value{
a matrix in sparseMatrix format, where negative values have been replaced by 0
}
\description{
This function converts regular (dense) matrices to sparse, which decreases overall RAM allocation.

It acts as a helper function to both \code{\link[=calc_distance_score_matrix]{calc_distance_score_matrix()}} and \code{\link[=calc_exact_score_matrix]{calc_exact_score_matrix()}}.

Since it is used in the score matching procedure, negative values are replaced by 0 to further reduce RAM usage.
}
\keyword{internal}
