% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{default_matching_fields}
\alias{default_matching_fields}
\title{Default matching fields for \code{\link[=biblioverlap]{biblioverlap()}}}
\format{
An object of class \code{list} of length 5.
}
\usage{
default_matching_fields
}
\description{
Each element in the named list corresponds to a column that will be used to match documents between bibliographic datasets.
}
\details{
In total, there are five columns that need to be specified. The default values come from column names in \href{https://www.lens.org/}{The Lens}.
}
\section{Fields}{

\describe{
\item{\code{doi}}{Unique identifier (e.g. DOI). Default: 'DOI'}

\item{\code{title}}{Document title. Default: 'Title'}

\item{\code{pubyear}}{Publication year. Default: 'Publication Year'}

\item{\code{authors}}{Authors. Default: 'Author/s'}

\item{\code{source}}{Publication source. Default: 'Source Title'}
}}

\keyword{internal}
