% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-doi_matching.R
\name{doi_matching}
\alias{doi_matching}
\title{Document matching using unique identifiers for two bibliographic datasets}
\usage{
doi_matching(db1, db2, n_threads, doi_score = 2)
}
\arguments{
\item{db1}{\itemize{
\item First bibliographic database in the comparison
}}

\item{db2}{\itemize{
\item Second bibliographic database in the comparison
}}

\item{n_threads}{\itemize{
\item number of (logical) cores to be used in the matching
}}

\item{doi_score}{\itemize{
\item custom score value to help identify DOI matches in previous comparisons
}}
}
\value{
a list containing match data
}
\description{
This function receives two lists of unique identifiers (e.g. DOI - Digital Object Identifier) and returns a list containing the matches of list1 against list2
This step of the matching is also given a "fake score" that will be used as a tag to identify and exclude matched documents from subsequent doi matching steps, increasing overall efficiency of the whole process
}
