% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-biblioverlap.R
\name{get_matching_summary_df}
\alias{get_matching_summary_df}
\title{Obtaining a table summary of biblioverlap's matching}
\usage{
get_matching_summary_df(matched_db_list)
}
\arguments{
\item{matched_db_list}{\itemize{
\item list of matched dataframes (with UUID column added by biblioverlap)
}}
}
\value{
a table summary the matching results
}
\description{
Obtaining a table summary of biblioverlap's matching
}
