% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-biblioverlap.R
\name{merge_results}
\alias{merge_results}
\title{Merge biblioverlap's results into a single dataframe}
\usage{
merge_results(db_list, filter_distinct = FALSE)
}
\arguments{
\item{db_list}{\itemize{
\item list of matched dataframes (with UUID column added by \code{\link{biblioverlap}})
}}

\item{filter_distinct}{\itemize{
\item boolean value determining whether to return only the subset containing distinct records (TRUE) or to keep overlapping records between datasets (FALSE). Default: FALSE
}}
}
\value{
a single dataframe containing data from db_list, featuring an additional 'SET_NAME' column to indicate from which dataset each record came
}
\description{
Merge biblioverlap's results into a single dataframe
}
\examples{

#Running document-level matching procedure for two datasets
biblioverlap_results <- biblioverlap(ufrj_bio_0122[1:2])

#Obtaining the results as a single dataframe (including overlapping records)
all_data <- merge_results(biblioverlap_results$db_list)

#Checking number of total rows and overlapping documents are in the dataframe
nrow(all_data)
sum(duplicated(all_data$UUID))

#Obtaining only unique records as a single dataframe
distinct_data <- merge_results(biblioverlap_results$db_list, filter_distinct = FALSE)

#Checking number of total rows and overlapping documents are in the dataframe
nrow(distinct_data)
sum(duplicated(distinct_data$UUID))

}
