% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plots_and_app.R
\name{plot_matching_summary}
\alias{plot_matching_summary}
\title{Plotting biblioverlap's matching summary}
\usage{
plot_matching_summary(matching_summary_df, ...)
}
\arguments{
\item{matching_summary_df}{\itemize{
\item summary of matching process generated by \code{\link[=biblioverlap]{biblioverlap()}}
}}

\item{...}{\itemize{
\item additional arguments passed down to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}
}}
}
\value{
a barplot summary of the matching results
}
\description{
Plotting biblioverlap's matching summary
}
\examples{
#Running document-level matching procedure
biblioverlap_results <- biblioverlap(ufrj_bio_0122)

#Checking biblioverlap results (summary table)
biblioverlap_results$summary

#Plotting the matching summary
plot_matching_summary(biblioverlap_results$summary)

}
