% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-bibliometrix_parsing.R
\name{plot_matching_summary}
\alias{plot_matching_summary}
\title{Obtaining a plot summary of biblioverlap's matching}
\usage{
plot_matching_summary(matching_summary_df, ...)
}
\arguments{
\item{matching_summary_df}{- summary of matching process generated by `matching_summary_df()`}

\item{...}{- additional arguments passed down to `ggplot2::geom_text()`}
}
\value{
a barplot summary of the matching results
}
\description{
Obtaining a plot summary of biblioverlap's matching
}
