% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plots_and_app.R
\name{plot_upset}
\alias{plot_upset}
\title{Plotting UpSet plot from biblioverlap results}
\usage{
plot_upset(db_list, ...)
}
\arguments{
\item{db_list}{\itemize{
\item list of matched dataframes (with UUID column added by biblioverlap)
}}

\item{...}{\itemize{
\item arguments to be passed down to \code{\link[UpSetR:upset]{UpSetR::upset()}}
}}
}
\value{
a UpSet plot
}
\description{
Plotting UpSet plot from biblioverlap results
}
