% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-bibliometrix_parsing.R
\name{plot_venn}
\alias{plot_venn}
\title{Plotting Venn Diagram from biblioverlap results}
\usage{
plot_venn(db_list, ...)
}
\arguments{
\item{db_list}{\itemize{
\item list of matched dataframes (with UUID column added by biblioverlap)
}}

\item{...}{\itemize{
\item Additional arguments that can be passed down to \code{\link[ggVennDiagram:ggVennDiagram]{ggVennDiagram::ggVennDiagram()}}
}}
}
\value{
a Venn Diagram
}
\description{
Plotting Venn Diagram from biblioverlap results
}
