% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-bibliometrix_parsing.R
\name{plot_venn}
\alias{plot_venn}
\title{Plotting Venn Diagram from biblioverlap results}
\usage{
plot_venn(db_list)
}
\arguments{
\item{db_list}{- list of matched dataframes (with UUID column added by biblioverlap)}
}
\value{
a Venn Diagram
}
\description{
Plotting Venn Diagram from biblioverlap results
}
