% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\encoding{UTF-8}
\name{ufrj_bio_0122}
\alias{ufrj_bio_0122}
\title{ufrj_bio_0122}
\format{
\subsection{\code{ufrj_bio_0122}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
ufrj_bio_0122
}
\description{
Data obtained from \href{https://www.lens.org}{The Lens Scholarly Search}
Report ...
}
\keyword{datasets}
